+++ app/src/main/java/com/huelvadevelopers/proyectozero/model/BankAccount.kt
@@ -6,6 +6,6 @@ import java.io.FileDescriptor
  * Created by DrAP on 07/07/2017.
  */
 
-class BankAccount (val id: Int, var name: String, var description: String, var balance: Int, val currency: String){
+class BankAccount (val id: Int, var name: String, var description: String, var balance: Float, val currency: String){
 
 }
